refactor: Make assignment sql driver a standalone crate
Split out the assignment sql driver to the standalone crate
