+++ crates/assignment-sql/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "openstack-keystone-assignment-sql"
+description = "OpenStack Keystone assignment SQL driver"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+exclude.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+openstack-keystone-core = { version = "0.1", path = "../core" }
+sea-orm.workspace = true
+tokio = { workspace = true, features = ["macros"] }
+tracing.workspace = true
+
+[dev-dependencies]
+openstack-keystone-core = { version = "0.1", path = "../core", features = ["mock"] }
+sea-orm = { workspace = true, features = ["mock", "sqlx-sqlite" ]}
+
+[lints]
+workspace = true
