+++ crates/assignment-sql/src/entity.rs
@@ -11,5 +11,11 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
+#![allow(clippy::doc_paragraphs_missing_punctuation)]
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.4
 
-pub use openstack_keystone_core::assignment::backend::AssignmentBackend;
+pub mod prelude;
+
+pub mod assignment;
+pub mod sea_orm_active_enums;
+pub mod system_assignment;
