+++ crates/assignment-sql/src/entity/assignment.rs
