+++ crates/assignment-sql/src/entity/prelude.rs
@@ -12,4 +12,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-pub use openstack_keystone_core::assignment::backend::AssignmentBackend;
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.4
+#![allow(unused_imports)]
+
+pub use super::assignment::Entity as Assignment;
+pub use super::system_assignment::Entity as SystemAssignment;
