+++ crates/assignment-sql/src/entity/sea_orm_active_enums.rs
@@ -12,4 +12,19 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-pub use openstack_keystone_core::assignment::backend::AssignmentBackend;
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.4
+
+use sea_orm::entity::prelude::*;
+
+#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
+#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "type")]
+pub enum Type {
+    #[sea_orm(string_value = "GroupDomain")]
+    GroupDomain,
+    #[sea_orm(string_value = "GroupProject")]
+    GroupProject,
+    #[sea_orm(string_value = "UserDomain")]
+    UserDomain,
+    #[sea_orm(string_value = "UserProject")]
+    UserProject,
+}
