+++ crates/assignment-sql/src/entity/system_assignment.rs
@@ -32,19 +32,12 @@ pub struct Model {
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::role::Entity",
-        from = "Column::RoleId",
-        to = "super::role::Column::Id"
-    )]
-    Role,
-}
+pub enum Relation {}
 
-impl Related<super::role::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Role.def()
-    }
-}
+//impl Related<super::role::Entity> for Entity {
+//    fn to() -> RelationDef {
+//        Relation::Role.def()
+//    }
+//}
 
 impl ActiveModelBehavior for ActiveModel {}
