+++ crates/core/src/assignment/error.rs
@@ -96,3 +96,12 @@ pub enum AssignmentProviderError {
         source: validator::ValidationErrors,
     },
 }
+
+impl From<crate::error::DatabaseError> for AssignmentProviderError {
+    fn from(source: crate::error::DatabaseError) -> Self {
+        match source {
+            cfl @ crate::error::DatabaseError::Conflict { .. } => Self::Conflict(cfl.to_string()),
+            other => Self::Driver(other.to_string()),
+        }
+    }
+}
