+++ crates/keystone/src/db/entity.rs
@@ -22,7 +22,6 @@ pub mod alembic_version;
 pub mod application_credential;
 pub mod application_credential_access_rule;
 pub mod application_credential_role;
-pub mod assignment;
 pub mod config_register;
 pub mod consumer;
 pub mod credential;
@@ -63,7 +62,6 @@ pub mod sea_orm_active_enums;
 pub mod sensitive_config;
 pub mod service;
 pub mod service_provider;
-pub mod system_assignment;
 pub mod token;
 pub mod token_restriction;
 pub mod token_restriction_role_association;
