+++ crates/keystone/src/db/entity/prelude.rs
@@ -21,7 +21,7 @@ pub use super::alembic_version::Entity as AlembicVersion;
 pub use super::application_credential::Entity as ApplicationCredential;
 pub use super::application_credential_access_rule::Entity as ApplicationCredentialAccessRule;
 pub use super::application_credential_role::Entity as ApplicationCredentialRole;
-pub use super::assignment::Entity as Assignment;
+//pub use super::assignment::Entity as Assignment;
 pub use super::config_register::Entity as ConfigRegister;
 pub use super::consumer::Entity as Consumer;
 pub use super::credential::Entity as Credential;
@@ -61,7 +61,7 @@ pub use super::role_option::Entity as RoleOption;
 pub use super::sensitive_config::Entity as SensitiveConfig;
 pub use super::service::Entity as Service;
 pub use super::service_provider::Entity as ServiceProvider;
-pub use super::system_assignment::Entity as SystemAssignment;
+//pub use super::system_assignment::Entity as SystemAssignment;
 pub use super::token::Entity as Token;
 pub use super::token_restriction::Entity as TokenRestriction;
 pub use super::token_restriction_role_association::Entity as TokenRestrictionRoleAssociation;
